refactor: cooldown system and consolidate kit/spawn config
- updated `cooldownManager.ts` to directly take durations explicitly.
- merged `spawnConfig.default.ts` and `kitsConfig.default.ts` into main `config.default.ts`.
- removed small configuration files for kits and spawn systems to avoid clutter.
- refactored teleport and kit managers to use `configManager.ts` instead.
- updated commands such as `/spawn`, `/rtp`, `/back`, `/warp` and `/home` to properly fetch configured durations when using `setCooldown`.
- updated `plan.md` state.
